+++ changelog.md
@@ -1,2 +1,9 @@
 # 0.1.0
 - Initial release
+# 0.2.0
+- Magic
+# 0.3.0
+- Added sub tree ignore via `<!-- i18n-checker:disable -->`
+# 0.4.0
+- Ignore URLs, Emails and trim punctuation
+- Added config file support via `--config`